+++ packages/website/index.html
@@ -8,7 +8,7 @@
     <script type="module" src="/src/main.tsx"></script>
   </head>
   <body>
-    <div id="app"></div>
+    <div id="app" class="flex min-h-full bg-white dark:bg-slate-900"></div>
     <div id="portal"></div>
   </body>
 </html>
