+++ packages/website/package.json
@@ -11,11 +11,14 @@
     "@nlxchat/react": "^2.0.0-beta.1",
     "@nlxchat/widget": "^2.0.0-beta.1",
     "htm": "^3.1.1",
+    "ramda": "^0.29.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.17.0",
     "tinycolor2": "^1.4.1"
   },
   "devDependencies": {
+    "@types/ramda": "^0.29.7",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@types/tinycolor2": "^1.4.2",
