+++ packages/website/src/custom-components/Disclaimer.css
@@ -1,0 +1,36 @@
+.chat-disclaimer {
+  border-radius: 6px;
+  padding: 6px;
+  color: #676767;
+  text-align: center;
+}
+
+.chat-disclaimer p,
+.chat-disclaimer a,
+.chat-disclaimer button {
+  font-size: 0.625rem;
+}
+
+.chat-disclaimer a {
+  text-decoration: underline;
+}
+
+.chat-disclaimer p {
+  margin-bottom: 6px;
+}
+
+.chat-disclaimer button {
+  border: none;
+  background: none;
+  color: var(--primaryColor);
+  cursor: pointer;
+  display: inline-block;
+}
+
+.chat-disclaimer button + button {
+  margin-left: 10px;
+}
+
+.chat-disclaimer button:hover {
+  color: var(--primaryColorLighter);
+}
