+++ packages/website/src/custom-components/FeedbackForm.css
@@ -1,0 +1,45 @@
+.chat-feedback-form {
+  padding: 10px;
+  border-radius: 6px;
+  background: #efefef;
+}
+
+.chat-feedback-form > * + * {
+  margin-top: 10px;
+}
+
+.chat-feedback-form input,
+.chat-feedback-form textarea {
+  display: block;
+  font-size: 0.75rem;
+  width: 100%;
+  padding: 6px;
+  background: #fff;
+  border: 1px solid #dedede;
+  border-radius: 6px;
+}
+
+.chat-feedback-form input:disabled,
+.chat-feedback-form textarea:disabled {
+  background-color: #efefef;
+  color: #676767;
+}
+
+.chat-feedback-form [type="submit"] {
+  display: block;
+  width: 100%;
+  background-color: var(--primaryColor);
+  padding: 8px 12px;
+  border-radius: 6px;
+  border: none;
+  color: #fff;
+  font-size: 0.75rem;
+  text-transform: uppercase;
+  font-weight: 600;
+  cursor: pointer;
+  letter-spacing: 0.05rem;
+}
+
+.chat-feedback-form [type="submit"]:hover {
+  background-color: var(--primaryColorLighter);
+}
