+++ packages/website/tailwind.config.js
@@ -2,7 +2,16 @@
 export default {
   content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
-    extend: {}
+    fontFamily: {
+      sans: ["'Inter'", "sans-serif"],
+      mono: ["'Space Mono'", "monospace"],
+      serif: ["Times", "serif"]
+    },
+    extend: {
+      maxWidth: {
+        "8xl": "88rem"
+      }
+    }
   },
   plugins: []
 };
